Refactor agent worktree context exports
