+++ Sources/RepoPrompt/Features/AgentMode/Views/Components/AgentOraclePill.swift
@@ -1,0 +1,337 @@
+import SwiftUI
+
+// MARK: - Oracle Pill
+
+enum AgentOraclePillLogic {
+    static func hasRenderableMessages(session: ChatSession, liveMessageCount: Int?) -> Bool {
+        if let liveMessageCount {
+            return liveMessageCount > 0
+        }
+        return session.hasMessages
+    }
+
+    static func eligibleSessions(
+        sessions: [ChatSession],
+        streamingSessionIDs: Set<UUID>,
+        liveMessageCount: (UUID) -> Int?,
+        activeAgentSessionID: UUID? = nil,
+        activeRunID: UUID? = nil
+    ) -> [ChatSession] {
+        let renderable = sessions.filter { session in
+            hasRenderableMessages(session: session, liveMessageCount: liveMessageCount(session.id))
+                || streamingSessionIDs.contains(session.id)
+        }
+        guard activeAgentSessionID != nil || activeRunID != nil else { return renderable }
+
+        func isUnownedLegacy(_ session: ChatSession) -> Bool {
+            session.agentModeSessionID == nil && session.agentModeRunID == nil
+        }
+        func matchesAgent(_ session: ChatSession) -> Bool {
+            guard let activeAgentSessionID else { return true }
+            return session.agentModeSessionID == activeAgentSessionID
+        }
+
+        if let activeRunID {
+            let exactRunMatches = renderable.filter { matchesAgent($0) && $0.agentModeRunID == activeRunID }
+            if !exactRunMatches.isEmpty { return exactRunMatches }
+
+            let sameAgentLegacyRunMatches = renderable.filter { matchesAgent($0) && $0.agentModeSessionID != nil && $0.agentModeRunID == nil }
+            if !sameAgentLegacyRunMatches.isEmpty { return sameAgentLegacyRunMatches }
+
+            if let activeAgentSessionID,
+               renderable.contains(where: { $0.agentModeSessionID == activeAgentSessionID })
+            {
+                return []
+            }
+            return renderable.filter(isUnownedLegacy)
+        }
+
+        if let activeAgentSessionID {
+            let sameAgentMatches = renderable.filter { $0.agentModeSessionID == activeAgentSessionID }
+            if !sameAgentMatches.isEmpty { return sameAgentMatches }
+        }
+
+        return renderable.filter(isUnownedLegacy)
+    }
+
+    static func latestSession(
+        in sessions: [ChatSession],
+        streamingSessionIDs: Set<UUID>
+    ) -> ChatSession? {
+        let streaming = sessions.filter { streamingSessionIDs.contains($0.id) }
+        if let latestStreaming = streaming.max(by: { $0.savedAt < $1.savedAt }) {
+            return latestStreaming
+        }
+        return sessions.max(by: { $0.savedAt < $1.savedAt })
+    }
+
+    static func selectedSessionID(
+        currentSelectionID: UUID?,
+        in sessions: [ChatSession],
+        streamingSessionIDs: Set<UUID>
+    ) -> UUID? {
+        if let currentSelectionID,
+           sessions.contains(where: { $0.id == currentSelectionID })
+        {
+            return currentSelectionID
+        }
+        return latestSession(in: sessions, streamingSessionIDs: streamingSessionIDs)?.id
+    }
+
+    static func reconciledPresentedSessionID(
+        currentSessionID: UUID?,
+        isExplicit: Bool,
+        sameTabSessions: [ChatSession],
+        eligibleSessions: [ChatSession],
+        streamingSessionIDs: Set<UUID>
+    ) -> UUID? {
+        if isExplicit {
+            guard let currentSessionID,
+                  sameTabSessions.contains(where: { $0.id == currentSessionID })
+            else {
+                return nil
+            }
+            return currentSessionID
+        }
+
+        if let currentSessionID,
+           eligibleSessions.contains(where: { $0.id == currentSessionID })
+        {
+            return currentSessionID
+        }
+        return latestSession(in: eligibleSessions, streamingSessionIDs: streamingSessionIDs)?.id
+    }
+
+    static func session(matchingChatID raw: String, in sessions: [ChatSession]) -> ChatSession? {
+        let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
+        guard !trimmed.isEmpty else { return nil }
+        return sessions.first { session in
+            session.id.uuidString == trimmed || session.shortID == trimmed
+        }
+    }
+}
+
+/// Pill that appears when there are oracle chat sessions for the current tab.
+/// More prominent when streaming. Clicking opens a wide popover with chat transcript.
+struct AgentOraclePill: View {
+    @ObservedObject var oracleViewModel: OracleViewModel
+    let windowID: Int
+    let currentTabID: UUID?
+    let activeAgentSessionID: UUID?
+    let activeRunID: UUID?
+
+    private enum PresentedSessionSource {
+        case latest
+        case explicit
+    }
+
+    @State private var showPopover = false
+    @State private var autoScrollEnabled = false
+    @State private var presentedSessionID: UUID?
+    @State private var presentedSessionSource: PresentedSessionSource = .latest
+    @ObservedObject private var fontScale = FontScaleManager.shared
+    private var fontPreset: FontScalePreset {
+        fontScale.preset
+    }
+
+    private var eligibleTabSessions: [ChatSession] {
+        guard let tabID = currentTabID else { return [] }
+        return AgentOraclePillLogic.eligibleSessions(
+            sessions: oracleViewModel.sessions(forTabID: tabID),
+            streamingSessionIDs: oracleViewModel.streamingSessions,
+            liveMessageCount: { oracleViewModel.liveMessageCount(for: $0) },
+            activeAgentSessionID: activeAgentSessionID,
+            activeRunID: activeRunID
+        )
+    }
+
+    private var latestTabSession: ChatSession? {
+        AgentOraclePillLogic.latestSession(
+            in: eligibleTabSessions,
+            streamingSessionIDs: oracleViewModel.streamingSessions
+        )
+    }
+
+    private var isStreaming: Bool {
+        guard let latestTabSession else { return false }
+        return oracleViewModel.streamingSessions.contains(latestTabSession.id)
+    }
+
+    private var presentedSession: ChatSession? {
+        guard let presentedSessionID,
+              let tabID = currentTabID else { return nil }
+        return oracleViewModel.sessions(forTabID: tabID).first { $0.id == presentedSessionID }
+    }
+
+    private var isPresentedSessionStreaming: Bool {
+        guard let presentedSessionID else { return isStreaming }
+        return oracleViewModel.streamingSessions.contains(presentedSessionID)
+    }
+
+    private var popoverSubtitle: String {
+        guard let presentedSession else { return "Latest tab chat" }
+        if presentedSession.id == latestTabSession?.id {
+            return "Latest tab chat"
+        }
+        return presentedSession.name
+    }
+
+    private var hasAnySessions: Bool {
+        latestTabSession != nil
+    }
+
+    var body: some View {
+        #if DEBUG
+            let _ = AgentModePerfDiagnostics.increment("ui.body.statusPills.oracle")
+        #endif
+        Group {
+            if hasAnySessions {
+                let cornerRadius = AgentPillMetrics.cornerRadius()
+                Button {
+                    openPopover(chatID: nil)
+                } label: {
+                    HStack(spacing: 6) {
+                        if isStreaming {
+                            ProgressView()
+                                .controlSize(.mini)
+                                .scaleEffect(0.7)
+                        } else {
+                            Image(systemName: "brain")
+                                .font(fontPreset.swiftUIFont(sizeAtNormal: 12))
+                                .foregroundStyle(.secondary)
+                        }
+                        Text("Oracle")
+                            .font(fontPreset.swiftUIFont(sizeAtNormal: 12, weight: isStreaming ? .semibold : .medium))
+                            .foregroundStyle(isStreaming ? .primary : .secondary)
+                    }
+                    .padding(.horizontal, AgentPillMetrics.horizontalPadding())
+                    .frame(height: AgentPillMetrics.height())
+                    .background(isStreaming ? AnyShapeStyle(.ultraThinMaterial) : AnyShapeStyle(.ultraThinMaterial))
+                    .clipShape(RoundedRectangle(cornerRadius: cornerRadius, style: .continuous))
+                    .overlay(
+                        RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+                            .stroke(isStreaming ? Color.purple.opacity(0.4) : Color.secondary.opacity(0.15), lineWidth: isStreaming ? 1 : 0.5)
+                    )
+                    .shadow(color: isStreaming ? Color.purple.opacity(0.15) : .clear, radius: 4, y: 1)
+                }
+                .buttonStyle(.plain)
+                .hoverTooltip(isStreaming ? "Oracle is thinking — click to view the live chat" : "Open the latest Oracle chat for this tab", .top)
+                .animation(.easeInOut(duration: 0.2), value: isStreaming)
+            } else {
+                Color.clear.frame(width: 0, height: 0)
+            }
+        }
+        .onReceive(NotificationCenter.default.publisher(for: .showAgentOraclePopover)) { note in
+            guard let targetWindowID = note.userInfo?["windowID"] as? Int,
+                  targetWindowID == windowID else { return }
+
+            let requestedTabID: UUID? = {
+                if let tabID = note.userInfo?["tabID"] as? UUID { return tabID }
+                if let tabIDString = note.userInfo?["tabID"] as? String { return UUID(uuidString: tabIDString) }
+                return nil
+            }()
+            if let requestedTabID, requestedTabID != currentTabID { return }
+
+            let requestedChatID: String? = {
+                if let chatID = note.userInfo?["chatID"] as? String { return chatID }
+                if let chatID = note.userInfo?["chatID"] as? UUID { return chatID.uuidString }
+                return nil
+            }()
+
+            openPopover(chatID: requestedChatID)
+        }
+        .popover(isPresented: $showPopover, arrowEdge: .bottom) {
+            oraclePopoverContent
+        }
+        .onChange(of: currentTabID) { _, _ in
+            reconcilePresentedSession()
+        }
+        .onChange(of: activeAgentSessionID) { _, _ in
+            reconcilePresentedSession()
+        }
+        .onChange(of: activeRunID) { _, _ in
+            reconcilePresentedSession()
+        }
+    }
+
+    @ViewBuilder
+    private var oraclePopoverContent: some View {
+        // Popover dimensions scale so chat messages don't feel cramped at
+        // Larger/Extra Large. Width gets a tighter cap than height because the
+        // popover is anchored to the composer and we don't want it to spill
+        // beyond the window edges; the chat transcript area takes the rest.
+        let popoverWidth = fontPreset.scaledClamped(800, max: 1040)
+        let transcriptMinHeight = fontPreset.scaledClamped(350, max: 460)
+        let transcriptIdealHeight = fontPreset.scaledClamped(500, max: 660)
+        let transcriptMaxHeight = fontPreset.scaledClamped(600, max: 780)
+        VStack(alignment: .leading, spacing: 8) {
+            HStack(spacing: 6) {
+                Text("Oracle")
+                    .font(fontPreset.swiftUIFont(sizeAtNormal: 13, weight: .semibold))
+                if isPresentedSessionStreaming {
+                    ProgressView()
+                        .controlSize(.mini)
+                        .scaleEffect(0.7)
+                }
+                Spacer()
+                Text(popoverSubtitle)
+                    .font(fontPreset.swiftUIFont(sizeAtNormal: 11))
+                    .foregroundStyle(.secondary)
+                    .lineLimit(1)
+            }
+
+            ChatMessagesView(
+                viewModel: oracleViewModel,
+                autoScrollEnabled: $autoScrollEnabled,
+                bottomOcclusion: 0,
+                showsScrollControls: true,
+                autoScrollOnAppear: true,
+                sessionIDOverride: presentedSessionID
+            )
+            .frame(minHeight: transcriptMinHeight, idealHeight: transcriptIdealHeight, maxHeight: transcriptMaxHeight)
+            .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
+        }
+        .padding(14)
+        .frame(width: popoverWidth)
+    }
+
+    private func reconcilePresentedSession() {
+        guard showPopover else { return }
+        let sameTabSessions = currentTabID.map { oracleViewModel.sessions(forTabID: $0) } ?? []
+        let resolvedID = AgentOraclePillLogic.reconciledPresentedSessionID(
+            currentSessionID: presentedSessionID,
+            isExplicit: presentedSessionSource == .explicit,
+            sameTabSessions: sameTabSessions,
+            eligibleSessions: eligibleTabSessions,
+            streamingSessionIDs: oracleViewModel.streamingSessions
+        )
+        guard let resolvedID else {
+            presentedSessionID = nil
+            showPopover = false
+            return
+        }
+        presentedSessionID = resolvedID
+    }
+
+    private func openPopover(chatID: String?) {
+        guard let tabID = currentTabID else { return }
+        let target: ChatSession?
+        let source: PresentedSessionSource
+        if let chatID {
+            target = AgentOraclePillLogic.session(
+                matchingChatID: chatID,
+                in: oracleViewModel.sessions(forTabID: tabID)
+            )
+            source = .explicit
+            guard target != nil else { return }
+        } else {
+            target = latestTabSession
+            source = .latest
+        }
+        guard let target else { return }
+
+        presentedSessionID = target.id
+        presentedSessionSource = source
+        showPopover = true
+    }
+}
