+++ Sources/RepoPrompt/Features/AgentMode/Views/Components/AgentTogglePills.swift
@@ -1,0 +1,81 @@
+import SwiftUI
+
+// MARK: - Auto Edit Pill
+
+struct AgentAutoEditPill: View {
+    let isOn: Bool
+    let onToggle: () -> Void
+
+    @ObservedObject private var fontScale = FontScaleManager.shared
+    private var fontPreset: FontScalePreset {
+        fontScale.preset
+    }
+
+    private var tooltipText: String {
+        if isOn {
+            return "Auto Edit is on: apply_edits writes files immediately after the agent proposes them."
+        }
+        return "Auto Edit is off: apply_edits requires approval. If sandbox permissions still allow file edits, those changes bypass RepoPrompt review."
+    }
+
+    var body: some View {
+        let cornerRadius = AgentPillMetrics.cornerRadius()
+        let dotSize = fontPreset.scaledMetric(CGFloat(7))
+        Button(action: onToggle) {
+            HStack(spacing: 6) {
+                Circle()
+                    .fill(isOn ? Color.green : Color.secondary.opacity(0.35))
+                    .frame(width: dotSize, height: dotSize)
+                Text("Auto Edit")
+                    .font(fontPreset.swiftUIFont(sizeAtNormal: 12, weight: .medium))
+                    .foregroundStyle(isOn ? Color.green : .secondary)
+            }
+            .padding(.horizontal, AgentPillMetrics.horizontalPadding())
+            .frame(height: AgentPillMetrics.height())
+            .background(.ultraThinMaterial)
+            .clipShape(RoundedRectangle(cornerRadius: cornerRadius, style: .continuous))
+            .overlay(
+                RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+                    .stroke(isOn ? Color.green.opacity(0.35) : Color.secondary.opacity(0.15), lineWidth: isOn ? 0.8 : 0.5)
+            )
+        }
+        .buttonStyle(.plain)
+        .hoverTooltip(tooltipText, .top)
+    }
+}
+
+// MARK: - Interview Pill
+
+struct AgentInterviewPill: View {
+    let isOn: Bool
+    let onToggle: () -> Void
+
+    @ObservedObject private var fontScale = FontScaleManager.shared
+    private var fontPreset: FontScalePreset {
+        fontScale.preset
+    }
+
+    var body: some View {
+        let cornerRadius = AgentPillMetrics.cornerRadius()
+        Button(action: onToggle) {
+            HStack(spacing: 5) {
+                Image(systemName: "questionmark.bubble")
+                    .font(fontPreset.swiftUIFont(sizeAtNormal: 11))
+                    .foregroundStyle(isOn ? Color.accentColor : .secondary)
+                Text("Interview")
+                    .font(fontPreset.swiftUIFont(sizeAtNormal: 12, weight: .medium))
+                    .foregroundStyle(isOn ? Color.accentColor : .secondary)
+            }
+            .padding(.horizontal, AgentPillMetrics.horizontalPadding())
+            .frame(height: AgentPillMetrics.height())
+            .background(.ultraThinMaterial)
+            .clipShape(RoundedRectangle(cornerRadius: cornerRadius, style: .continuous))
+            .overlay(
+                RoundedRectangle(cornerRadius: cornerRadius, style: .continuous)
+                    .stroke(isOn ? Color.accentColor.opacity(0.4) : Color.secondary.opacity(0.15), lineWidth: isOn ? 0.8 : 0.5)
+            )
+        }
+        .buttonStyle(.plain)
+        .hoverTooltip(isOn ? "Interview is on: the agent will ask clarifying questions before starting" : "Interview is off: the agent will start working immediately", .top)
+    }
+}
