+++ Tests/RepoPromptTests/Prompt/PromptContextPreAssemblyServiceTests.swift
@@ -1,0 +1,213 @@
+@testable import RepoPrompt
+import XCTest
+
+final class PromptContextPreAssemblyServiceTests: XCTestCase {
+    private actor CapturedPaths {
+        private var value: [String] = []
+        func set(_ paths: [String]) {
+            value = paths
+        }
+
+        func get() -> [String] {
+            value
+        }
+    }
+
+    private var temporaryRoots = FileSystemTemporaryRoots()
+
+    override func tearDownWithError() throws {
+        temporaryRoots.removeAll()
+        try super.tearDownWithError()
+    }
+
+    func testResolveUsesWorktreeContentAndLogicalizesFileTree() async throws {
+        let fixture = try await makeBoundFixture()
+        let request = PromptContextPreAssemblyRequest(
+            cfg: makeConfig(gitInclusion: .complete),
+            selection: StoredSelection(selectedPaths: ["Sources/App.swift"], codemapAutoEnabled: false),
+            store: fixture.store,
+            lookupContext: fixture.lookupContext,
+            filePathDisplay: .full,
+            onlyIncludeRootsWithSelectedFiles: true,
+            showCodeMapMarkers: true,
+            selectedGitDiffFolderPolicy: .filesOnly,
+            selectedGitDiffProvider: { _ in "unexpected selected diff" },
+            completeGitDiffProvider: { "base checkout complete diff must not appear" }
+        )
+
+        let result = await PromptContextPreAssemblyService.resolve(request)
+
+        XCTAssertEqual(result.physicalSelection.selectedPaths, [fixture.worktreeRoot.appendingPathComponent("Sources/App.swift").standardizedFileURL.path])
+        XCTAssertEqual(result.entries.count, 1)
+        XCTAssertTrue(result.entries.first?.loadedContent?.contains("worktree") ?? false)
+        XCTAssertFalse(result.entries.first?.loadedContent?.contains("base") ?? true)
+        XCTAssertTrue(result.fileTreeContent?.contains(fixture.logicalRoot.standardizedFileURL.path) ?? false, result.fileTreeContent ?? "")
+        XCTAssertFalse(result.fileTreeContent?.contains(fixture.worktreeRoot.standardizedFileURL.path) ?? true, result.fileTreeContent ?? "")
+        XCTAssertEqual(result.gitDiff, PromptContextGitDiffPolicy.deferredCompleteWorktreeGitDiffMessage)
+    }
+
+    func testResolveSelectedDiffUsesPhysicalizedSelectionAndPolicy() async throws {
+        let fixture = try await makeBoundFixture()
+        let captured = CapturedPaths()
+        let request = PromptContextPreAssemblyRequest(
+            cfg: makeConfig(gitInclusion: .selected),
+            selection: StoredSelection(selectedPaths: ["Sources"], codemapAutoEnabled: false),
+            store: fixture.store,
+            lookupContext: fixture.lookupContext,
+            filePathDisplay: .relative,
+            onlyIncludeRootsWithSelectedFiles: true,
+            showCodeMapMarkers: true,
+            selectedGitDiffFolderPolicy: .expandFolders,
+            selectedGitDiffLookupProfile: .uiAssisted,
+            selectedGitDiffProvider: { paths in
+                await captured.set(paths)
+                return "selected diff"
+            },
+            completeGitDiffProvider: { "unexpected complete diff" }
+        )
+
+        let result = await PromptContextPreAssemblyService.resolve(request)
+        let paths = await captured.get()
+
+        XCTAssertEqual(result.gitDiff, "selected diff")
+        XCTAssertEqual(Set(paths), Set([
+            fixture.worktreeRoot.appendingPathComponent("Sources/App.swift").standardizedFileURL.path,
+            fixture.worktreeRoot.appendingPathComponent("Sources/Keep.swift").standardizedFileURL.path
+        ]))
+        XCTAssertFalse(paths.contains(fixture.logicalRoot.appendingPathComponent("Sources/App.swift").standardizedFileURL.path))
+    }
+
+    func testSelectedDiffArtifactPolicyCanRespectGitInclusionNone() async throws {
+        let root = try makeTemporaryRoot(name: "PromptPreAssemblyDiffArtifact")
+        let diffText = "diff --git a/Sources/App.swift b/Sources/App.swift\n"
+        try FileSystemTestSupport.write(diffText, to: root.appendingPathComponent("_git_data/diff/selected.diff"))
+        let store = WorkspaceFileContextStore()
+        _ = try await store.loadRoot(path: root.path)
+        let selection = StoredSelection(selectedPaths: ["_git_data/diff/selected.diff"], codemapAutoEnabled: false)
+        let baseRequest = PromptContextPreAssemblyRequest(
+            cfg: makeConfig(gitInclusion: .none),
+            selection: selection,
+            store: store,
+            lookupContext: WorkspaceLookupContext(rootScope: .allLoaded, bindingProjection: nil),
+            filePathDisplay: .relative,
+            onlyIncludeRootsWithSelectedFiles: true,
+            showCodeMapMarkers: true,
+            selectedGitDiffFolderPolicy: .expandFolders,
+            selectedGitDiffProvider: { _ in "unexpected selected provider" },
+            completeGitDiffProvider: { "unexpected complete provider" }
+        )
+        let includeResult = await PromptContextPreAssemblyService.resolve(baseRequest)
+
+        let respectRequest = PromptContextPreAssemblyRequest(
+            cfg: makeConfig(gitInclusion: .none),
+            selection: selection,
+            store: store,
+            lookupContext: WorkspaceLookupContext(rootScope: .allLoaded, bindingProjection: nil),
+            filePathDisplay: .relative,
+            onlyIncludeRootsWithSelectedFiles: true,
+            showCodeMapMarkers: true,
+            selectedGitDiffFolderPolicy: .expandFolders,
+            selectedGitDiffArtifactPolicy: .respectGitInclusion,
+            selectedGitDiffProvider: { _ in "unexpected selected provider" },
+            completeGitDiffProvider: { "unexpected complete provider" }
+        )
+        let respectResult = await PromptContextPreAssemblyService.resolve(respectRequest)
+
+        let includeClipboard = await PromptPackagingService.generateClipboardContent(
+            metaInstructions: [],
+            userInstructions: "",
+            files: includeResult.entries,
+            fileTreeContent: includeResult.fileTreeContent,
+            gitDiff: includeResult.gitDiff,
+            includeSavedPrompts: false,
+            includeFiles: true,
+            includeUserPrompt: false,
+            filePathDisplay: .relative,
+            codemapSnapshots: includeResult.codemapSnapshots,
+            promptSectionsOrder: PromptAssemblyBuilder.defaultSectionOrder,
+            disabledPromptSections: [],
+            duplicateUserInstructionsAtTop: false
+        )
+        let respectClipboard = await PromptPackagingService.generateClipboardContent(
+            metaInstructions: [],
+            userInstructions: "",
+            files: respectResult.entries,
+            fileTreeContent: respectResult.fileTreeContent,
+            gitDiff: respectResult.gitDiff,
+            includeSavedPrompts: false,
+            includeFiles: true,
+            includeUserPrompt: false,
+            filePathDisplay: .relative,
+            codemapSnapshots: respectResult.codemapSnapshots,
+            promptSectionsOrder: PromptAssemblyBuilder.defaultSectionOrder,
+            disabledPromptSections: [],
+            duplicateUserInstructionsAtTop: false
+        )
+
+        XCTAssertEqual(includeResult.gitDiff, diffText)
+        XCTAssertTrue(includeClipboard.contains(diffText), includeClipboard)
+        XCTAssertNil(respectResult.gitDiff)
+        XCTAssertTrue(respectResult.entries.isEmpty)
+        XCTAssertFalse(respectClipboard.contains("<git_diff>"), respectClipboard)
+        XCTAssertFalse(respectClipboard.contains(diffText), respectClipboard)
+    }
+
+    private func makeBoundFixture() async throws -> (
+        logicalRoot: URL,
+        worktreeRoot: URL,
+        store: WorkspaceFileContextStore,
+        lookupContext: WorkspaceLookupContext
+    ) {
+        let logicalRoot = try makeTemporaryRoot(name: "PromptPreAssemblyLogical")
+        let worktreeRoot = try makeTemporaryRoot(name: "PromptPreAssemblyWorktree")
+        try FileSystemTestSupport.write("let origin = \"base\"\n", to: logicalRoot.appendingPathComponent("Sources/App.swift"))
+        try FileSystemTestSupport.write("let origin = \"keep-base\"\n", to: logicalRoot.appendingPathComponent("Sources/Keep.swift"))
+        try FileSystemTestSupport.write("let origin = \"worktree\"\n", to: worktreeRoot.appendingPathComponent("Sources/App.swift"))
+        try FileSystemTestSupport.write("let origin = \"keep-worktree\"\n", to: worktreeRoot.appendingPathComponent("Sources/Keep.swift"))
+
+        let store = WorkspaceFileContextStore()
+        _ = try await store.loadRoot(path: logicalRoot.path)
+        let sessionID = UUID()
+        let binding = makeBinding(logicalRoot: logicalRoot, worktreeRoot: worktreeRoot)
+        let materializedProjection = await WorkspaceRootBindingProjectionMaterializer(store: store).materialize(sessionID: sessionID, bindings: [binding])
+        let projection = try XCTUnwrap(materializedProjection)
+        let lookupContext = WorkspaceLookupContext(rootScope: projection.lookupRootScope, bindingProjection: projection)
+        return (logicalRoot, worktreeRoot, store, lookupContext)
+    }
+
+    private func makeConfig(gitInclusion: GitInclusion) -> PromptContextResolved {
+        PromptContextResolved(
+            includeFiles: true,
+            includeUserPrompt: true,
+            includeMetaPrompts: false,
+            includeFileTree: true,
+            fileTreeMode: .auto,
+            codeMapUsage: .none,
+            gitInclusion: gitInclusion,
+            storedPromptIds: []
+        )
+    }
+
+    private func makeBinding(logicalRoot: URL, worktreeRoot: URL) -> AgentSessionWorktreeBinding {
+        AgentSessionWorktreeBinding(
+            id: "bind_test",
+            repositoryID: "repo_test",
+            repoKey: "repo",
+            logicalRootPath: logicalRoot.path,
+            logicalRootName: logicalRoot.lastPathComponent,
+            worktreeID: "worktree_test",
+            worktreeRootPath: worktreeRoot.path,
+            worktreeName: worktreeRoot.lastPathComponent,
+            branch: "feature/test",
+            head: "abcdef",
+            visualLabel: "test",
+            visualColorHex: "#3366FF",
+            boundAt: Date(timeIntervalSinceReferenceDate: 123),
+            source: "test"
+        )
+    }
+
+    private func makeTemporaryRoot(name: String) throws -> URL {
+        try temporaryRoots.makeRoot(suiteName: name)
+    }
+}
